Update data-model.js
Created additional math functions for sections 1-5 (up to finances section).  

Other notes:  

I changed the app function because when I tried to test it I couldn't get it to run correctly.  

I also had a difficult time using functions that stored arrays locally, so instead I made the arrays global, put info inside them and then used them for future calculations.

Regarding the last point, I always had an intuition that this was not the best way to do it, but couldn't figure out an easier way to get it done so I kept doing it.  However, I did the formulas for calculating price last, and I used a slightly different method that didn't require me to keep making new global arrays.  Would be interested in hearing your opinions on how to do this.  

I realize some of the changes I made to your code might not be optimal, but I wanted to be able to work on formulas that I could get to work and have you critique them.